requirements.txt
seaborn
pandas
matplotlib
tabulate
pycryptodome